fixed both unit tests

+++ tests/pass/koala_max_topological_cut.cpp
@@ -28,11 +28,11 @@ TEST(MaxTopologicalCut, lessThanCMXMemory)
     mv::ControlModel cm(om);
 
     auto output = cm.getOutput();
-    int maxTopologicalCutValue;
+    uint64_t maxTopologicalCutValue;
 
     /*Get the max topological cut value*/
     if(output->hasAttr("MaxTopologicalCutValue"))
-        maxTopologicalCutValue = output->get<int>("MaxTopologicalCutValue");
+        maxTopologicalCutValue = output->get<uint64_t>("MaxTopologicalCutValue");
 
     /*The max topological cut of the equivalent network in the PoC compiler is 492032*/
     ASSERT_EQ(maxTopologicalCutValue, 492032);
